Refactor agent_cli and git_commit_ai: envvar, inline, merge tests
- agent_cli/cli.py: use Typer envvar= for OPENAI_MODEL and
  SYSTEM_INSTRUCTIONS instead of manual os.getenv(); drop os import
  and module-level DEFAULT_MODEL/SYSTEM_INSTRUCTIONS constants
- git_commit_ai/cli.py: inline _run_editor_flow into commit(); it was
  only called once and adds indirection without benefit
- git_commit_ai: merge test_amend.py and test_stage_all.py into
  test_cli.py (both tested cli.py utilities); fix duplicate __main__
  block; update BUILD.bazel accordingly
